// webapp/src/shared/generated/models/GamePairingStatusDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * transcendence-app
 * The transcendence-app API description
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface GamePairingStatusDto
 */
export interface GamePairingStatusDto {
    /**
     * 
     * @type {string}
     * @memberof GamePairingStatusDto
     */
    gameRoomId: string | null;
    /**
     * 
     * @type {string}
     * @memberof GamePairingStatusDto
     */
    gameQueueStatus: GamePairingStatusDtoGameQueueStatusEnum;
}


/**
 * @export
 */
export const GamePairingStatusDtoGameQueueStatusEnum = {
    None: 'none',
    Playing: 'playing',
    Waiting: 'waiting'
} as const;
export type GamePairingStatusDtoGameQueueStatusEnum = typeof GamePairingStatusDtoGameQueueStatusEnum[keyof typeof GamePairingStatusDtoGameQueueStatusEnum];


/**
 * Check if a given object implements the GamePairingStatusDto interface.
 */
export function instanceOfGamePairingStatusDto(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "gameRoomId" in value;
    isInstance = isInstance && "gameQueueStatus" in value;

    return isInstance;
}

export function GamePairingStatusDtoFromJSON(json: any): GamePairingStatusDto {
    return GamePairingStatusDtoFromJSONTyped(json, false);
}

export function GamePairingStatusDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): GamePairingStatusDto {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'gameRoomId': json['gameRoomId'],
        'gameQueueStatus': json['gameQueueStatus'],
    };
}

export function GamePairingStatusDtoToJSON(value?: GamePairingStatusDto | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'gameRoomId': value.gameRoomId,
        'gameQueueStatus': value.gameQueueStatus,
    };
}